# API_documentation/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/administrator/modules": {
      "get": {
        "operationId": "ModuleController_getModules",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Here are all the Modules available"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get modules",
        "tags": [
          "System Management"
        ]
      }
    },
    "/administrator/module/{id}": {
      "get": {
        "operationId": "ModuleController_getModule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the module with submodules"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get module by ID",
        "tags": [
          "System Management"
        ]
      },
      "put": {
        "operationId": "ModuleController_updateModule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Payload to update the module info",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Module updated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update current module",
        "tags": [
          "System Management"
        ]
      }
    },
    "/administrator/module": {
      "post": {
        "operationId": "ModuleController_createModule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Module created successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "409": {
            "description": "Conflict - Resource already exist or duplicate entry"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new Module",
        "tags": [
          "System Management"
        ]
      }
    },
...
    "/audit": {
      "get": {
        "operationId": "AuditController_getAuditLogs",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "resource",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "success",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get audit logs with filters",
        "tags": [
          "Administrator"
        ]
      }
    },
    "/audit/resource/{resource}/{id}": {
      "get": {
        "operationId": "AuditController_getResourceHistory",
        "parameters": [
          {
            "name": "resource",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get audit history fo ra specific resource",
        "tags": [
          "Administrator"
        ]
      }
    },
    "/audit/user/{id}/activity": {
      "get": {
        "operationId": "AuditController_getUserActivity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get user activity report",
        "tags": [
          "Administrator"
        ]
      }
    },
    "/audit/my-activity": {
      "get": {
        "operationId": "AuditController_getMyActivity",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get own activity report",
        "tags": [
          "Administrator"
        ]
      }
    },
...
  },
  "info": {
    "title": "ABAS v3 API",
    "description": "API documentation for ABAS v3 project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      },
      "security_clearance": {
        "type": "apiKey",
        "name": "Security-Clearance",
        "in": "header",
        "description": "Required security clearance level (documentational only)."
      }
    },
    "schemas": {
...
      "CreateModuleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Operations",
            "description": "The name of the module"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateModuleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "New module name",
            "description": "Update current module name"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}